+++ backloggd-plus.meta.js
@@ -10,7 +10,7 @@
 // @name:ko           Backloggd Plus
 // @name:pl           Backloggd Plus
 // @namespace         https://github.com/NemoKing1210/backloggd-plus
-// @version           0.7.4
+// @version           0.7.6
 // @description       Extends Backloggd and adds a Backloggd button on Steam game pages
 // @description:ru    Расширяет Backloggd и добавляет кнопку Backloggd на страницах игр Steam
 // @description:zh-CN 扩展 Backloggd：更多游戏信息、更丰富的界面与使用体验
@@ -45,6 +45,8 @@
 // @connect            gamestatus.info
 // @connect            howlongtobeat.com
 // @connect            api.opencritic.com
+// @connect            opencritic.com
+// @connect            html.duckduckgo.com
 // @connect            www.protondb.com
 // @connect            protondb.com
 // @run-at             document-idle